Create project with `cargo new minimap`
minimap will be a command line program that takes source code file(s) as
input and produces an image containing the minimap(s) of the code, as
seen in many text editors these days.

It will also take diffs as input, and produce an image showing the
minimap of each changed file with the added and removed lines
highlighted. This will be useful for visualizing git commits or PRs, and
for showing an overview of changes to be made in project-based
tutorials.